% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveperm.R
\docType{package}
\name{adaptiveperm-package}
\alias{adaptiveperm}
\alias{adaptiveperm-package}
\title{adaptiveperm}
\description{
\code{adaptiveperm} is a package for adaptive permutation testing
}
\author{
\strong{Maintainer}: Timothy Barry \email{tbarry@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-4356-627X}{ORCID})

Authors:
\itemize{
  \item Lasse Fischer
  \item Aaditya Ramdas
}

}
\keyword{package}
