% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{evaluate_simulation_results}
\alias{evaluate_simulation_results}
\title{Evaluate simulation results}
\usage{
evaluate_simulation_results(result_df, under_null)
}
\arguments{
\item{result_df}{the results data frame outputted by \code{run_permutation_test()}}

\item{under_null}{a logical vector indicating whether a given hypothesis is under the null (\code{TRUE}) or under the alternative (\code{FALSE})}
}
\value{
a vector containing the false discovery proportion (FDP) and true positive proportion (TPP)
}
\description{
\code{evaluate_simulation_results()} evaluates the results of \code{run_permutation_test()} when run on simulated data.
}
\examples{
n <- 50L
m <- 500L
x <- rbinom(n = n, size = 1, prob = 0.5)
under_null <- sample(c(rep(TRUE, 0.9 * m), rep(FALSE, 0.1 * m)))
Y_list <- sapply(X = seq_len(m), FUN = function(i) {
 mus <- x * (if (under_null[i]) 0 else 1)
 rnorm(n = n, mean = mus)
}, simplify = FALSE)

# MW statistic
res <- run_permutation_test(Y_list, x, test_statistic = "MW")
evaluate_simulation_results(res, under_null)

# sum over treated units statistic
res <- run_permutation_test(Y_list, x, test_statistic = "sum_over_treated_units")
evaluate_simulation_results(res, under_null)
}
