% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_adaptive_perm_test.R
\name{run_permutation_test}
\alias{run_permutation_test}
\title{Run adaptive permutation test}
\usage{
run_permutation_test(
  Y_list,
  x,
  side = "two_tailed",
  h = 15L,
  alpha = 0.1,
  test_statistic = "MW",
  adaptive = TRUE
)
}
\value{
the results data frame
}
\description{
Run adaptive permutation test
}
\examples{
n <- 50L
m <- 500L
x <- rbinom(n = n, size = 1, prob = 0.5)
under_null <- sample(c(rep(TRUE, 0.9 * m), rep(FALSE, 0.1 * m)))
Y_list <- sapply(X = seq_len(m), FUN = function(i) {
 mus <- x * (if (under_null[i]) 0 else 1)
 rnorm(n = n, mean = mus)
}, simplify = FALSE)
res <- run_permutation_test(Y_list, x)

# evaluate results

}
